module_generator/requirements.txt
pyyaml
argparse
termcolor
